refactor(domain): 나트륨 필드 타입 변경 및 영양소 필드 확장

+++ src/main/java/com/jdc/recipe_service/domain/dto/recipe/RecipeNutritionDto.java
@@ -23,5 +23,5 @@ public class RecipeNutritionDto {
     private BigDecimal sugar;
 
     @Schema(description = "나트륨 (mg)")
-    private Integer sodium;
+    private BigDecimal sodium;
 }
